+++ src/groups/faction_warfare.rs
@@ -8,8 +8,8 @@ pub struct FactionWarfareGroup<'a> {
 #[derive(Debug, Deserialize)]
 #[allow(missing_docs)]
 pub struct FactionLeaderboardItem {
-    pub amount: Option<u64>,
-    pub faction_id: Option<u64>,
+    pub amount: Option<i32>,
+    pub faction_id: Option<i32>,
 }
 
 #[derive(Debug, Deserialize)]
@@ -30,8 +30,8 @@ pub struct FWLeaderboards {
 #[derive(Debug, Deserialize)]
 #[allow(missing_docs)]
 pub struct CharacterLeaderboardItem {
-    pub amount: u64,
-    pub character_id: u64,
+    pub amount: i32,
+    pub character_id: i32,
 }
 
 #[derive(Debug, Deserialize)]
@@ -52,8 +52,8 @@ pub struct FWCharacterLeaderboards {
 #[derive(Debug, Deserialize)]
 #[allow(missing_docs)]
 pub struct CorporationLeaderboardItem {
-    pub amount: u64,
-    pub corporation_id: u64,
+    pub amount: i32,
+    pub corporation_id: i32,
 }
 
 #[derive(Debug, Deserialize)]
@@ -74,9 +74,9 @@ pub struct FWCorporationLeaderboards {
 #[derive(Debug, Deserialize)]
 #[allow(missing_docs)]
 pub struct FWStatsItemRange {
-    pub total: u32,
-    pub last_week: u32,
-    pub yesterday: u32,
+    pub total: i32,
+    pub last_week: i32,
+    pub yesterday: i32,
 }
 
 #[derive(Debug, Deserialize)]
@@ -103,8 +103,8 @@ pub struct FWSystem {
 #[derive(Debug, Deserialize)]
 #[allow(missing_docs)]
 pub struct FWWar {
-    pub faction_id: u32,
-    pub against_id: u32,
+    pub faction_id: i32,
+    pub against_id: i32,
 }
 
 impl<'a> FactionWarfareGroup<'a> {
